updates to vis.config
