+++ app/package.json
@@ -33,7 +33,7 @@
     "remark-math": "^6.0.0",
     "socket.io-client": "^4.8.1",
     "three": "^0.165.0",
-    "znsocket": "^0.2.5"
+    "znsocket": "^0.2.6"
   },
   "devDependencies": {
     "@types/react": "^18.3.12",
