+++ tests/test_analysis.py
@@ -1,5 +1,16 @@
 from zndraw import ZnDraw
+import znsocket
 
+
+def run_queue(vis, key, msg: dict):
+    modifier_queue = znsocket.Dict(
+        r=vis.r,
+        socket=vis._refresh_client,
+        key=f"queue:{vis.token}:{key}",
+    )
+    modifier_queue.update(msg)
+    vis.socket.emit("room:worker:run")
+    vis.socket.sleep(5)
 
 def test_run_analysis_distance(server, s22_energy_forces):
     vis = ZnDraw(url=server, token="test_token")
@@ -7,8 +18,8 @@ def test_run_analysis_distance(server, s22_energy_forces):
     assert vis.figures == {}
     vis.selection = [0, 1]
 
-    vis.socket.emit("analysis:run", {"method": {"discriminator": "Distance"}})
-    vis.socket.sleep(7)
+    run_queue(vis, "analysis", {"Distance": {}})
+
     fig = vis.figures["Distance"]
     # assert that the x-axis label is "step"
     assert fig.layout.xaxis.title.text == "step"
@@ -18,10 +29,9 @@ def test_run_analysis_Properties1D(server, s22_energy_forces):
     vis = ZnDraw(url=server, token="test_token")
     vis.extend(s22_energy_forces)
     assert vis.figures == {}
-    vis.socket.emit(
-        "analysis:run", {"method": {"discriminator": "Properties1D", "value": "energy"}}
-    )
-    vis.socket.sleep(7)
+
+    run_queue(vis, "analysis", {"Properties1D": {"value": "energy"}})
+
     fig = vis.figures["Properties1D"]
     # assert that the x-axis label is "step"
     assert fig.layout.xaxis.title.text == "step"
@@ -32,18 +42,9 @@ def test_run_analysis_Properties2D(server, s22_energy_forces):
     vis = ZnDraw(url=server, token="test_token")
     vis.extend(s22_energy_forces)
     assert vis.figures == {}
-    vis.socket.emit(
-        "analysis:run",
-        {
-            "method": {
-                "discriminator": "Properties2D",
-                "x_data": "energy",
-                "y_data": "step",
-                "color": "energy",
-            }
-        },
-    )
-    vis.socket.sleep(7)
+
+    run_queue(vis, "analysis", {"Properties2D": {"x_data": "energy", "y_data": "step", "color": "energy"}})
+
     fig = vis.figures["Properties2D"]
     # assert that the x-axis label is "step"
     assert fig.layout.yaxis.title.text == "step"
@@ -56,8 +57,8 @@ def test_run_analysis_DihedralAngle(server, s22_energy_forces):
     assert vis.figures == {}
     vis.selection = [0, 1, 3, 4]
 
-    vis.socket.emit("analysis:run", {"method": {"discriminator": "DihedralAngle"}})
-    vis.socket.sleep(7)
+    run_queue(vis, "analysis", {"DihedralAngle": {}})
+    
     fig = vis.figures["DihedralAngle"]
     # assert that the x-axis label is "step"
     assert fig.layout.xaxis.title.text == "step"
