+++ tests/test_utils.py
@@ -1,15 +1,10 @@
-from unittest.mock import patch
-
-import ase
 import numpy as np
 import numpy.testing as npt
 
-from zndraw.scene import Scene
 from zndraw.utils import (
     convert_url_to_http,
     direction_to_euler,
     euler_to_direction,
-    get_schema_with_instance_defaults,
     parse_url,
 )
 
@@ -66,12 +61,3 @@ def test_url():
     url = convert_url_to_http(before_url)
     assert url == "https://localhost:8000/token/eNwsdW5k"
 
-
-def test_get_schema_with_instance_defaults():
-    scene = Scene(fps=33)
-    # patch Scene._get_atoms() to return a dummy Atoms object
-    with patch("zndraw.scene.Scene._get_atoms") as mock_get_atoms:
-        mock_get_atoms.return_value = ase.Atoms()
-        schema = get_schema_with_instance_defaults(scene)
-
-    assert schema["properties"]["fps"]["default"] == 33
